// homegroups/mobile/src/store/types.ts
import {SponsorshipState} from '../store/slices/sponsorshipSlice';
import {AuthState} from '../store/slices/authSlice';
import {GroupsState} from '../store/slices/groupsSlice';
import {TransactionsState} from '../store/slices/transactionsSlice';
import {TreasuryState} from '../store/slices/treasurySlice';
import {MeetingsState} from '../store/slices/meetingsSlice';
import {AnnouncementsState} from '../store/slices/announcementsSlice';
import {MembersState} from '../store/slices/membersSlice';
import {ChatState} from '../store/slices/chatSlice';
import {ServicePositionsState} from '../store/slices/servicePositionsSlice';
import {ReportsState} from '../store/slices/reportsSlice';
import {TreasurerHandoffState} from '../store/slices/treasurerHandoffSlice';
import {BusinessMeetingsState} from '../store/slices/businessMeetingsSlice';
import {DirectMessagesState} from '../store/slices/directMessagesSlice';
import {DashboardState} from './slices/dashboardSlice';
import {AdminRemovalState} from '../store/slices/adminRemovalSlice';
import {RecurringTransactionsState} from '../store/slices/recurringTransactionsSlice';
import {EngagementState} from '../store/slices/engagementSlice';
import {ReferralState} from '../store/slices/referralSlice';
import {StepWorkState} from '../store/slices/stepWorkSlice';
import {ReflectionsState} from '../store/slices/reflectionsSlice';
import {LiteratureState} from '../store/slices/literatureSlice';
import {GroupResourcesState} from '../store/slices/groupResourcesSlice';
import {GroupHealthState} from '../store/slices/groupHealthSlice';

export interface RootState {
  auth: AuthState;
  groups: GroupsState;
  transactions: TransactionsState;
  treasury: TreasuryState;
  meetings: MeetingsState;
  announcements: AnnouncementsState;
  members: MembersState;
  chat: ChatState;
  servicePositions: ServicePositionsState;
  sponsorship: SponsorshipState;
  reports: ReportsState;
  treasurerHandoff: TreasurerHandoffState;
  businessMeetings: BusinessMeetingsState;
  directMessages: DirectMessagesState;
  dashboard: DashboardState;
  adminRemoval: AdminRemovalState;
  recurringTransactions: RecurringTransactionsState;
  engagement: EngagementState;
  referral: ReferralState;
  stepWork: StepWorkState;
  reflections: ReflectionsState;
  literature: LiteratureState;
  groupResources: GroupResourcesState;
  groupHealth: GroupHealthState;
}